Create trait that will be used for a more optimized implementation
The original plan was to do multiple passes over the input text, with each pass
handling the insertion of a specific special character, such as em dashes. This
results in a time complexity of nm, where n is the length of the input text in
characters and m is the number of passes.

We can optimize this by doing a _single_ pass over in the input text, and on each
character, we run a set of "formatters" that may or may not do text manipulation
depending on the current character. Each formatter handles a separate special
character and keeps track of its own state independent of the others.

With this implementation, the upper bound on time complexity is nm, where
this time m is the number of formatters. However, in practice, the runtime
will be significantly less than nm because we expect only 1 formatter at most
to run on each character, and all other formatters will skip.