Restructure formatter files for clarity
The formatters will be used in other modules, so their module has been
renamed to "formatters" to make it clear that this where the formatters
can be found.

The "formatters" module has also been put into a parent
"text_manipulation" module. This is because the next module to be
written will handle combining the formatters and should be grouped
together with the formatters, but it isn't a formatter itself.