Fix consecutive formatter preventing next formatters from running
If the current count of consecutive matching characters is more than 1,
but the current input character is non-matching, the consecutive
characters formatter would return all of the consecutive characters that
it matched so that they don't get erased in the output.

However, because the previous implementation behaved as
"formatter returns any characters" = "formatter found a full match",
this caused subsequent formatters to not be run even when they
should be matching. Example: "Hi." <-- The " will not be replaced.

The solution is to make each formatter return 2 pieces of info:
- Any characters that should be added to the overall output
- A boolean indicating whether the next formatter should run or not
This allows formatters to return characters for the output while also
letting subsequent formatters run.