Remove double newlines from input files
In Markdown, it's common to use double spacing to separate each
paragraph. However, these double newlines should be replaced when
converting to a manuscript, as the paragraph separation will instead be
handled via indentation that can be applied in applications such as
LibreOffice Writer.

For now, we can reasonably assume most input files won't have more than
two newlines in a row. But in the future, we can implement an
"UnlimitedConsecutiveCharFormatter" that would fix the rare mistake of
having more than two newlines in a row.